+++ .storybook/preview.tsx
@@ -38,6 +38,15 @@ addDecorator((StoryFn) => (
 export const parameters = {
   // https://storybook.js.org/docs/react/essentials/actions#automatically-matching-args
   actions: { argTypesRegex: "^on.*" },
+  backgrounds: {
+    default: "Light mode page background",
+    values: [
+      { name: "Light mode page background", value: "#FFFFFF" },
+      { name: "Light mode default background", value: "#F5F5F5" },
+      { name: "Dark mode page background", value: "#121212" },
+      { name: "Dark mode default background", value: "#191919" },
+    ],
+  },
   // Load the Reservoir's Chakra-based theme.
   chakra: {
     theme: nyplTheme,